+++ lib/src/builders/channel.dart
@@ -58,15 +58,15 @@ class Channel<P> {
 }
 
 // TODO: these entry class are not particularly useful.  Consider removing them
-// class _ChannelEntry<P, M> {
-//   final Channel<P> channel;
-//   final FutureOr<P> Function(MessageContext ctx, M message)? payload;
-//   _ChannelEntry(this.channel, this.payload);
+class _ChannelEntry<P, M> {
+  final Channel<P> channel;
+  final FutureOr<P> Function(MessageContext ctx, M message)? payload;
+  _ChannelEntry(this.channel, this.payload);
 
-//   void enter(MessageHandlerBuilder<M> builder, bool reenterTarget) {
-//     builder.goTo(channel.to, payload: payload, reenterTarget: reenterTarget);
-//   }
-//}
+  void enter(MessageHandlerBuilder<M> builder, bool reenterTarget) {
+    builder.goTo(channel.to, payload: payload, reenterTarget: reenterTarget);
+  }
+}
 
 class _ChannelEntryWithData<P, M, D> {
   final Channel<P> channel;