+++ lib/tree_builders.dart
@@ -29,5 +29,3 @@ part 'src/builders/handlers/messages/goto_descriptor.dart';
 part 'src/builders/handlers/messages/goto_self_descriptor.dart';
 part 'src/builders/handlers/messages/stay_or_unhandled_descriptor.dart';
 part 'src/builders/handlers/messages/when_descriptor.dart';
-
-final _log = Logger('tree_state_machine.Builders');